fix to read real image size
attribute size of NSImage did not always contain the real image size but returns a size information that is screen resolution dependent and those value can be fractional.
To get the real image size it is needed to iterate over the representations array of type NSImageRep and find the largest size in this array.